# openapi.json
{
...
  "info": {
    "description": "Auto-generated route inventory for this iterion instance. Routes and methods are exact; request/response schemas are populated for the typed surface (see components) and enriched route-by-route. Note: the /api/v1/native, /api/v1/dispatcher and /api/v1/mcp/board sub-trees are served but registered on a separate mux, so they are intentionally not catalogued here.",
    "title": "iterion API",
    "version": "current"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/local/skills/{name}": {
      "delete": {
        "operationId": "deleteLocalSkillsByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "DELETE /api/local/skills/{name}",
        "tags": [
          "local"
        ]
      },
      "get": {
        "operationId": "getLocalSkillsByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/local/skills/{name}",
        "tags": [
          "local"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "name",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "put": {
        "operationId": "putLocalSkillsByName",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "PUT /api/local/skills/{name}",
        "tags": [
          "local"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}